+++ src/features/survey/questionList/QuestionWithDnd.tsx
@@ -54,10 +54,15 @@ function QuestionWithDnd({ item, dragRef, onIsDrag, offIsDrag, onDelete }: Props
   };
 
   const onItemPointerUp = (e: React.PointerEvent<SVGSVGElement>) => {
+    console.warn('onItemPointerUp: ');
     const target = e.target as HTMLDivElement;
+    console.warn('dragRef: ', dragRef);
 
     if (dragRef && checkIsInner(dragRef, target)) {
+      console.warn('delete', item.title);
       onDelete(item.title);
+    } else {
+      console.warn('not delete');
     }
   };
 
@@ -88,7 +93,23 @@ function QuestionWithDnd({ item, dragRef, onIsDrag, offIsDrag, onDelete }: Props
       <Question
         item={item}
         rightElement={
-          <MenuIcon onPointerDown={(e) => dragControls.start(e)} onPointerUp={onItemPointerUp} css={menuIconCss} />
+          <MenuIcon
+            onPointerDown={(e) => dragControls.start(e)}
+            onDrop={(e) => {
+              console.warn('onDrop: ');
+              const target = e.target as HTMLDivElement;
+              console.warn('dragRef: ', dragRef);
+
+              if (dragRef && checkIsInner(dragRef, target)) {
+                console.warn('delete', item.title);
+                onDelete(item.title);
+              } else {
+                console.warn('not delete');
+              }
+            }}
+            onPointerUp={onItemPointerUp}
+            css={menuIconCss}
+          />
         }
       />
     </Reorder.Item>
@@ -115,6 +136,7 @@ const checkIsInner = (dragRef: React.RefObject<HTMLDivElement>, target: HTMLDivE
   const targetRect = target.getBoundingClientRect();
 
   const isInner = dragRect.y < targetRect.y && targetRect.y < dragRect.y + dragRect.height;
+  console.warn('y: ', dragRect.y, targetRect.y);
 
   return isInner;
 };
